Remove all changes except for atomic_cas overloads
